Transfer_Money_Service/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
pydantic==2.10.2
pydantic_core==2.27.1
python-decouple==3.8
sqlparse==0.5.2
typing_extensions==4.12.2
